Moving threads to a separate package
