+++ src/main/java/org/openhab/binding/synologysurveillancestation/handler/SynologySurveillanceStationHandler.java
@@ -28,8 +28,8 @@ import org.eclipse.smarthome.core.thing.binding.builder.ThingBuilder;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
 import org.eclipse.smarthome.core.types.State;
-import org.openhab.binding.synologysurveillancestation.internal.SynoApiThreadEvent;
-import org.openhab.binding.synologysurveillancestation.internal.SynoApiThreadSnapshot;
+import org.openhab.binding.synologysurveillancestation.internal.thread.SynoApiThreadEvent;
+import org.openhab.binding.synologysurveillancestation.internal.thread.SynoApiThreadSnapshot;
 import org.openhab.binding.synologysurveillancestation.internal.webapi.SynoEvent;
 import org.openhab.binding.synologysurveillancestation.internal.webapi.WebApiException;
 import org.slf4j.Logger;
