+++ src/main/java/org/openhab/binding/synologysurveillancestation/internal/thread/SynoApiThread.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.synologysurveillancestation.internal;
+package org.openhab.binding.synologysurveillancestation.internal.thread;
 
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
